Move `formatDate` to `helpers/dates`
