+++ frontend/src/components/chats/MessagesDateGroups.tsx
@@ -3,6 +3,7 @@ import dayjs from "dayjs"
 
 import { useMessages } from "@/api/chats/messages"
 import Message from "@/components/chats/Message"
+import { formatDate } from "@/helpers/dates"
 
 const MessagesDateGroups: FC<{ messageIds: number[]; showStickyDate: boolean }> = ({
   messageIds,
@@ -56,21 +57,6 @@ const MessagesDateGroups: FC<{ messageIds: number[]; showStickyDate: boolean }> 
       ))}
     </>
   )
-}
-
-const formatDate = (timestamp: string) => {
-  const date = dayjs(timestamp)
-  const today = dayjs()
-
-  if (date.format("YYYY-MM-DD") === today.format("YYYY-MM-DD")) {
-    return "Today"
-  }
-
-  if (date.add(1, "day").format("YYYY-MM-DD") === today.format("YYYY-MM-DD")) {
-    return "Yesterday"
-  }
-
-  return date.format(date.year() === dayjs().year() ? "MMMM D" : "MMMM D, YYYY")
 }
 
 export default memo(MessagesDateGroups)
